refactor: LInt声明与实现分离
